add credit and debit

+++ src/main/java/rinhaw/com/example/rinha/infraestructure/controllers/transacoes/dto/TransacaoOutputDTO.java
@@ -1,6 +1,5 @@
 package rinhaw.com.example.rinha.infraestructure.controllers.transacoes.dto;
 
-import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TransacaoOutputDTO {
-   BigDecimal saldo;
-   BigDecimal limite;
+   long saldo;
+   long limite;
 }
